+++ Sources/HIGThemesContract/HIGTheme+Defaults.swift
@@ -77,6 +77,7 @@ extension HIGTheme {
     public var barChart: any HIGBarChartTokens { HIGSystemBarChartTokens() }
     public var lineChart: any HIGLineChartTokens { HIGSystemLineChartTokens() }
     public var pieChart: any HIGPieChartTokens { HIGSystemPieChartTokens() }
+    public var areaChart: any HIGAreaChartTokens { HIGSystemAreaChartTokens() }
     public var photoPicker: any HIGPhotoPickerTokens { HIGSystemPhotoPickerTokens() }
     public var photoEditor: any HIGPhotoEditorTokens { HIGSystemPhotoEditorTokens() }
     public var longTextEditor: any HIGLongTextEditorTokens { HIGSystemLongTextEditorTokens() }